+++ src/main/resources/db/migration/V1_2__table_transactions.sql
@@ -2,7 +2,7 @@ CREATE TABLE transactions (
   id            INTEGER PRIMARY KEY,
   block_id      INTEGER NULLABLE REFERENCES blocks,
   hash          VARCHAR NOT NULL UNIQUE,
-  amount        INTEGER NOT NULL,
+  amount        BIGINT NOT NULL,
   timestamp     BIGINT  NOT NULL,
   recipient_key VARCHAR NOT NULL,
   sender_key    VARCHAR NOT NULL,
